Move read_config to glados.configs instead
